src/requirements.txt
amqp==2.6.0
billiard==3.6.3.0
celery==4.4.0
Django==2.2.9
django-cors-headers==3.2.0
django-coreapi==2.3.0
django-environ==0.4.5
django-filter==2.2.0
django-redis==4.11.0
djangorestframework==3.11.0
#gunicorn==20.0.4
kombu==4.6.11
psycopg2-binary==2.8.4
pytz==2020.1
sqlparse==0.3.1
vine==1.3.0
coolname==1.1.0
#gevent==1.4.0
uwsgi